+++ src/models/User.js
@@ -399,17 +399,16 @@ module.exports = function( db, sequelize, DataTypes ) {
 				if (userId) {
 					return {
 						where: sequelize.or(
-							{ userId },
+							{ id: userId },
 							{ viewableByRole: 'all' },
-							{ viewableByRole: null },
-							{ viewableByRole: roles[userRole] || '' },
+							{ viewableByRole: roles[userRole] || 'all' },
 						)
 					};
 				} else {
 					return {
 						where: sequelize.or(
 							{ viewableByRole: 'all' },
-							{ viewableByRole: roles[userRole] || '' },
+							{ viewableByRole: roles[userRole]  || 'all' },
 						)
 					};
 				}