+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove unused global vimeo fields
 * Added possibility to query GET overview routes with NOT (!=) and substring (like%%) operator
 * Add option to list user data publically or per role
+* Allow viewablebyRole to be edited by the user
 
 ## v0.8.0 (2020-11-02)
 * Add id & extraData to properties included in idea GET call with param includeUser=1